test: share IPv6 availability helper
