+++ tests/test_main.py
@@ -1,6 +1,5 @@
 import importlib
 import inspect
-import socket
 import sys
 from logging import WARNING
 from pathlib import Path
@@ -9,7 +8,7 @@ import httpx
 import pytest
 
 import uvicorn.server
-from tests.utils import run_server
+from tests.utils import has_ipv6, run_server
 from uvicorn import Server
 from uvicorn._types import ASGIReceiveCallable, ASGISendCallable, Scope
 from uvicorn.config import Config
@@ -25,21 +24,6 @@ async def app(scope: Scope, receive: ASGIReceiveCallable, send: ASGISendCallable
     await send({"type": "http.response.body", "body": b"", "more_body": False})
 
 
-def _has_ipv6(host: str):
-    sock = None
-    has_ipv6 = False
-    if socket.has_ipv6:
-        try:
-            sock = socket.socket(socket.AF_INET6)
-            sock.bind((host, 0))
-            has_ipv6 = True
-        except Exception:  # pragma: no cover
-            pass
-    if sock:
-        sock.close()
-    return has_ipv6
-
-
 @pytest.mark.parametrize(
     "host, url",
     [
@@ -49,7 +33,7 @@ def _has_ipv6(host: str):
             "::1",
             "http://[::1]",
             id="ipv6",
-            marks=pytest.mark.skipif(not _has_ipv6("::1"), reason="IPV6 not enabled"),
+            marks=pytest.mark.skipif(not has_ipv6("::1"), reason="IPV6 not enabled"),
         ),
     ],
 )
